+++ ext/duckdb/duckdb.c
@@ -57,11 +57,11 @@ Init_duckdb_native(void) {
     rbduckdb_init_duckdb_instance_cache();
     rbduckdb_init_duckdb_value_impl();
     rbduckdb_init_duckdb_scalar_function();
-    rbduckdb_init_duckdb_init_info();
     rbduckdb_init_duckdb_function_info();
     rbduckdb_init_duckdb_vector();
     rbduckdb_init_duckdb_data_chunk();
     rbduckdb_init_memory_helper();
     rbduckdb_init_duckdb_table_function();
     rbduckdb_init_duckdb_bind_info();
+    rbduckdb_init_duckdb_init_info();
 }