+++ ext/duckdb/table_function_init_info.h
@@ -1,5 +1,5 @@
-#ifndef RUBY_DUCKDB_INIT_INFO_H
-#define RUBY_DUCKDB_INIT_INFO_H
+#ifndef RUBY_DUCKDB_TABLE_FUNCTION_INIT_INFO_H
+#define RUBY_DUCKDB_TABLE_FUNCTION_INIT_INFO_H
 
 struct _rubyDuckDBInitInfo {
     duckdb_init_info info;
@@ -7,6 +7,7 @@ struct _rubyDuckDBInitInfo {
 
 typedef struct _rubyDuckDBInitInfo rubyDuckDBInitInfo;
 
+extern VALUE cDuckDBTableFunctionInitInfo;
 rubyDuckDBInitInfo *get_struct_init_info(VALUE obj);
 void rbduckdb_init_duckdb_init_info(void);